LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-try-typed-om-expected.txt


FAIL Effects of position-try-fallbacks are visible in the computed values assert_equals: expected 30 but got 9999
